debug with ta's and got front end showing
