+++ client/package.json
@@ -2,12 +2,17 @@
   "name": "client",
   "version": "0.1.0",
   "private": true,
+  "proxy": "http://localhost.3001",
   "dependencies": {
+    "@apollo/client": "^3.7.3",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^13.5.0",
+    "graphql": "^16.6.0",
+    "jwt-decode": "^3.1.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-router-dom": "^6.6.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
