+++ client/src/App.js
@@ -1,16 +1,9 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import {
-  ApolloClient,
-  InMemoryCache,
-  ApolloProvider,
-  createHttpLink,
-} from "@apollo/client";
+
 import logo from "./logo.svg";
 
-//ask ta if these two are in correct location and if setcontext is needed. It is not in the mini project
 import "./App.css";
-import { setContext } from "@apollo/client/link/context";
 
 import Home from "./pages/Home";
 import Profile from "./pages/Profile";
@@ -20,60 +13,17 @@ import Login from "./pages/Login";
 // import Scheduling from "./pages/Scheduling";
 // import Services from "./pages/Services";
 
-const authLink = setContext((_, { headers }) => {
-  const token = localStorage.getItem("id_token");
-  return {
-    headers: {
-      ...headers,
-      authorization: token ? `Bearer ${token}` : "",
-    },
-  };
-});
-
-// Construct our main GraphQL API endpoint
-const httpLink = createHttpLink({
-  uri: "/graphql",
-});
-const client = new ApolloClient({
-  link: authLink.concat(httpLink),
-  cache: new InMemoryCache(),
-});
-
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        {/* <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a> */}
-      </header>
-      <ApolloProvider client={client}>
-        <Router>
-          <div className="flex-column justify-flex-start min-100-vh">
-            <Header />
-            <div className="container">
-              <Routes>
-                <Route path="/" element={<Home />} />
-                <Route path="/login" element={<Login />} />
-                <Route path="/signup" element={<Signup />} />
-                <Route path="/profile/:profileId" element={<Profile />} />
-                {/* <Route path="/services" element={<Services />} />
-                <Route path="/scheduling" element={<Scheduling />} /> */}
-              </Routes>
-            </div>
-            <Footer />
-          </div>
-        </Router>
-      </ApolloProvider>
+    <div>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/signup" element={<Signup />} />
+        <Route path="/profile/:profileId" element={<Profile />} />
+        {/* <Route path="/services" element={<Services />} />
+        <Route path="/scheduling" element={<Scheduling />} /> */}
+      </Routes>
     </div>
   );
 }
