+++ client/src/pages/Profile.js
@@ -1,12 +1,12 @@
-import React from 'react';
+import React from "react";
 
-import { useParams } from 'react-router-dom';
-import { useQuery } from '@apollo/client';
+import { useParams } from "react-router-dom";
+import { useQuery } from "@apollo/client";
 
-import SkillsList from '../components/SkillsList';
-import SkillForm from '../components/SkillForm';
+// import SkillsList from '../components/SkillsList';
+// import SkillForm from '../components/SkillForm';
 
-import { QUERY_SINGLE_PROFILE } from '../utils/queries';
+import { QUERY_SINGLE_PROFILE } from "../utils/queries";
 
 const Profile = () => {
   const { profileId } = useParams();
@@ -20,19 +20,7 @@ const Profile = () => {
   if (loading) {
     return <div>Loading...</div>;
   }
-  return (
-    // <div>
-    //   <h2 className="card-header">
-    //     {profile.name}'s friends have endorsed these skills...
-    //   </h2>
-    //
-    //   {profile.skills?.length > 0 && <SkillsList skills={profile.skills} />}
-    //
-    //   <div className="my-4 p-4" style={{ border: '1px dotted #1a1a1a' }}>
-    //     <SkillForm profileId={profile._id} />
-    //   </div>
-    // </div>
-  );
+  return;
 };
 
 export default Profile;
