+++ client/src/utils/mutations.js
@@ -2,7 +2,7 @@ import { gql } from "@apollo/client";
 
 export const ADD_PROFILE = gql`
   mutation addProfile($name: String!, $email: String!, $password: String!) {
-    // addProfile(name: $name, email: $email, password: $password pets: $pets) {
+    addProfile(name: $name, email: $email, password: $password) {
       token
       profile {
         _id
